+++ packages/react-email/src/utils/get-emails-directory-metadata.ts
@@ -1,4 +1,3 @@
-'use server';
 /* eslint-disable @typescript-eslint/no-non-null-assertion */
 import fs from 'node:fs';
 import path from 'node:path';