+++ packages/create-email/template/CHANGELOG.md
@@ -1,5 +1,17 @@
 # react-email-starter
 
+## 0.1.8
+
+### Patch Changes
+
+- react-email@3.0.6
+
+## 0.1.7
+
+### Patch Changes
+
+- @react-email/components@0.0.32
+
 ## 0.1.6
 
 ### Patch Changes